+++ public/index.html
@@ -38,7 +38,7 @@
         height: 100%;
         font-family: "Righteous", cursive;
       }
-      .modal.root {
+      .modal-root {
         position: fixed;
         z-index: 100;
         left: 0;
