+++ src/components/MarsPlot.tsx
@@ -6,21 +6,13 @@ import Header from "./Header";
 
 function MarsPlot() {
   const text = "Tomter";
-  const [modalOpen, setOpenModal] = useState(false);
-  const closeModal = () => setOpenModal(false);
-  const openModal = () => setOpenModal(true);
+
   return (
     <>
       <Header text={text}></Header>
 
       <div style={cardsContainer}>
-        {modalOpen && (
-          <Modal closeModal={closeModal}>
-            <p>HEj på dig MOdal</p>
-            <button onClick={closeModal}>close Modal</button>
-          </Modal>
-        )}
-        <CardContainer modalOpen={openModal} />
+        <CardContainer />
       </div>
       {/* <div style={mainContainer}> */}
       {/* <div className="cardContainer" style={cardContainerStyle}> */}
